refactor(components): organize each component into its own folder
