+++ packages/lets-ui-components/src/components/card/card.scss
