+++ packages/lets-ui-components/src/components/divider/divider.ts
