+++ packages/lets-ui-components/src/components/dropdown-menu/dropdown-menu.scss
