+++ packages/lets-ui-components/src/components/shortcut/shortcut.ts
