+++ packages/lets-ui-components/src/components/switch/switch.scss
