+++ packages/lets-ui-components/src/components/tabs/tabs.ts
