+++ packages/lets-ui-components/src/components/textarea/textarea.scss
